# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Finance Dashboard System API
  description: Backend API for the Finance Dashboard System
  version: 1.0.0
  contact:
    name: Finance Dashboard Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.finance-dashboard.example.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FinancialRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
...
paths:
...
  /records:
    post:
      tags:
        - Financial Records
      summary: Create a financial record
      description: Create a new income or expense record. Admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  format: double
                type:
                  type: string
                  enum: [income, expense]
                category:
                  type: string
                date:
                  type: string
                  format: date
                notes:
                  type: string
                  nullable: true
              required:
                - amount
                - type
                - category
                - date
            example:
              amount: 5000
              type: income
              category: Salary
              date: 2026-04-05
              notes: Monthly salary payment

      responses:
        201:
          description: Record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialRecord'
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        401:
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Insufficient permissions (must be Admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Financial Records
      summary: List financial records
      description: Retrieve a paginated list of financial records. Analyst and Admin only.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number (1-indexed)
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Records per page
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
          description: Filter by record type
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Filter records on or after this date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: Filter records on or before this date
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [date, createdAt]
            default: date
          description: Sort field
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort direction

      responses:
        200:
          description: List of records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FinancialRecord'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
                      hasNextPage:
                        type: boolean
                      hasPreviousPage:
                        type: boolean
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        401:
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Insufficient permissions (must be Analyst or Admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /records/{id}:
    get:
      tags:
        - Financial Records
      summary: Get a specific financial record
      description: Retrieve a single record by ID. Analyst and Admin only.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Record details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialRecord'
        401:
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Insufficient permissions (must be Analyst or Admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Financial Records
      summary: Update a financial record
      description: Update record details. Admin only.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  format: double
                type:
                  type: string
                  enum: [income, expense]
                category:
                  type: string
                date:
                  type: string
                  format: date
                notes:
                  type: string
                  nullable: true
            example:
              amount: 5100
              notes: Updated salary payment

      responses:
        200:
          description: Record updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialRecord'
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        401:
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Insufficient permissions (must be Admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Financial Records
      summary: Delete a financial record
      description: Permanently delete a record. Admin only.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Record deleted successfully
        401:
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Insufficient permissions (must be Admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...